[BD-162] refactor: DTO 패키지 및 클래스명 리팩토링
- requestDto, responseDto 패키지 → request, response 패키지로 수정
- `UsernameCheckResponse` 수정 : UsernameCheckResponseDto로 클래스명 수정
