+++ api/src/main/java/com/example/api/domain/bucket/dto/response/UsernameCheckResponseDto.java
@@ -1,11 +1,11 @@
-package com.example.api.domain.bucket.dto.responseDto;
+package com.example.api.domain.bucket.dto.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
 @Getter
 @AllArgsConstructor
-public class UsernameCheckResponse {
+public class UsernameCheckResponseDto {
 
     private String message;
     private boolean available;
